+++ nequip/model/__init__.py
@@ -6,7 +6,11 @@ from .saved_models import (
     ModelFromPackage,
     ModelTypeNamesFromPackage,
 )
-from .nequip_models import NequIPGNNModel, FullNequIPGNNModel
+from .nequip_models import (
+    NequIPGNNModel,
+    PresetNequIPGNNModel,
+    FullNequIPGNNModel,
+)
 from .pair_potential import ZBLPairPotential
 
 __all__ = [
@@ -17,6 +21,7 @@ __all__ = [
     "ModelFromPackage",
     "ModelTypeNamesFromPackage",
     "NequIPGNNModel",
+    "PresetNequIPGNNModel",
     "FullNequIPGNNModel",
     "ZBLPairPotential",
 ]